+++ cmd/postgres_exporter/namespace.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/blang/semver"
-	"github.com/go-kit/kit/log/level"
+	"github.com/go-kit/log/level"
 	"github.com/lib/pq"
 	"github.com/prometheus/client_golang/prometheus"
 )