+++ cmd/postgres_exporter/queries.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	"github.com/blang/semver"
-	"github.com/go-kit/kit/log/level"
+	"github.com/go-kit/log/level"
 	"gopkg.in/yaml.v2"
 )
 
@@ -120,7 +120,7 @@ var queryOverrides = map[string][]OverrideQuery{
 
 	"pg_stat_archiver": {
 		{
-			semver.MustParseRange(">=0.0.0"),
+			semver.MustParseRange(">=9.4.0"),
 			`
 			SELECT *,
 				extract(epoch from now() - last_archived_time) AS last_archive_age